error correction

+++ tasks/seq/karaseva_e_num_of_alternations_signs/func_tests/main.cpp
@@ -18,21 +18,19 @@ TEST(karaseva_e_num_of_alternations_signs_seq, test_alternation) {
     in[i] = (i % 2 == 0) ? 1 : -1;
   }
 
-  // Create task_data
   auto task_data_seq = std::make_shared<ppc::core::TaskData>();
   task_data_seq->inputs.emplace_back(reinterpret_cast<uint8_t *>(in.data()));
   task_data_seq->inputs_count.emplace_back(in.size());
   task_data_seq->outputs.emplace_back(reinterpret_cast<uint8_t *>(out.data()));
   task_data_seq->outputs_count.emplace_back(out.size());
 
-  // Create Task
   karaseva_e_num_of_alternations_signs_seq::AlternatingSignsSequential test_task_sequential(task_data_seq);
   ASSERT_EQ(test_task_sequential.Validation(), true);
   test_task_sequential.PreProcessing();
   test_task_sequential.Run();
   test_task_sequential.PostProcessing();
 
-  EXPECT_EQ(out[0], kCount - 1);
+  EXPECT_EQ(static_cast<long>(out[0]), kCount - 1);  // Converting out[0] ê long
 }
 
 TEST(karaseva_e_num_of_alternations_signs_seq, test_alternations_every_two) {
@@ -57,7 +55,7 @@ TEST(karaseva_e_num_of_alternations_signs_seq, test_alternations_every_two) {
   test_task_sequential.Run();
   test_task_sequential.PostProcessing();
 
-  EXPECT_EQ(out[0], (kCount / 2) - 1);
+  EXPECT_EQ(static_cast<long>(out[0]), (kCount / 2) - 1);  // Converting out[0] ê long
 }
 
 TEST(karaseva_e_num_of_alternations_signs_seq, test_no_alternations) {
@@ -78,5 +76,5 @@ TEST(karaseva_e_num_of_alternations_signs_seq, test_no_alternations) {
   test_task_sequential.Run();
   test_task_sequential.PostProcessing();
 
-  EXPECT_EQ(out[0], 0);
+  EXPECT_EQ(static_cast<long>(out[0]), 0);  // Converting out[0] ê long ( if (lhs == rhs))
 }
